requirements.txt
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.4.0
thop==0.1.1-2209072238
tifffile==2021.8.8
timm==0.9.16
tokenizers==0.15.0
tomli==2.0.1
torchmetrics==1.2.0
torchstat==0.0.7
traitlets==5.11.2
transformers==4.36.2
triton==2.0.0
types-python-dateutil==2.8.19.14
typing-extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
wandb==0.15.12
wcwidth==0.2.12
websocket-client==1.6.4
websockets==11.0.3
wheel==0.42.0
yacs==0.1.8
yapf==0.31.0